+++ internal/vision/ip.go
@@ -1,0 +1,7 @@
+package vision
+
+import "net"
+
+func validateIP(ip string) bool {
+	return net.ParseIP(ip) != nil
+}